feat(daily-update-r2): Wave 5 (solo) — task 015 (FR-07 abstract dependencies + types/services hoist, zero solution-local imports)
Substantial scope-expansion beyond the bare FR-07 wording. To satisfy "zero solution-local imports in the new
package" the agent had to hoist additional artifacts left as back-pointers by Wave 3/4 tasks:

Hoisted to @spaarke/daily-briefing-components (4 new files, byte-identical):
- src/types/notifications.ts (323 LOC) — was at src/solutions/DailyBriefing/src/types/
- src/services/notificationService.ts (330 LOC)
- src/services/preferencesService.ts (176 LOC)
- src/utils/toastUtils.ts (15 LOC) + utils/index.ts barrel

Package changes:
- package.json — added @spaarke/auth ^2.0.0 peer dep (omitted by task 010 scaffold, legit per FR-07)
- src/index.ts — added export * from './utils'
- src/types/index.ts — added BriefingDependencies interface + AuthenticatedFetch type alias + re-exported notifications surface
- src/services/index.ts — added 7 new function exports
- briefingService.ts — switched to direct @spaarke/auth import; intra-package types path
- 5 hooks + 3 components — intra-package relative paths only

BriefingDependencies surface (FR-07 contract):
  { authenticatedFetch, webApi, userId, tenantId, onRateLimitError?, onRecordOpen? }

Re-export shims preserved at 4 original locations under src/solutions/DailyBriefing/ to keep pre-existing
consumers compiling: types/notifications, services/notificationService, services/preferencesService, utils/toastUtils.

Grep verification per FR-07:
  grep -rE "from\s+[\"'][^\"']*solutions/" src/client/shared/Spaarke.DailyBriefing.Components/src/
  → ZERO matches. Also zero Spaarke.Auth back-pointer paths. All "solutions/" text in new package is JSDoc-comment-only
  (intentional hoist-origin history).

ADR-024 BINDING preserved: TODO_REGARDING_CATALOG + applyResolverFields in useInlineTodoCreate unchanged at the
logic level — only type-import paths for IWebApi/NotificationItem/NotificationPriority were rewritten to resolve
through the hoisted types (same identifiers, same semantics).

tsc --noEmit on new package: 3 cross-package alias "errors" remain (@spaarke/ui-components, @spaarke/auth peer
declaration) — these are the standard shared-lib type-check pattern matching Smart Todo + Events Components.
Resolved at consumer's Vite bundle, not at package-local tsc. ZERO regressions, ZERO new back-pointer errors.

Forbidden zones honored: no BFF, no LegalWorkspace, no SpaarkeAi, no authInit.ts, no runtimeConfig.ts, no
MicrosoftToDoIcon, no TASK-INDEX.md (committed separately).

TASK-INDEX.md flipped 015 to ✅; status counter 19 → 20 of 36.
