+++ projects/spaarke-daily-update-service-r2/current-task.md
@@ -10,16 +10,21 @@
 
 | Field | Value |
 |-------|-------|
-| **Task** | 054 — Migrate LegalWorkspace + SpaarkeAi to auth factory; thin-wrap authInit.ts |
-| **Step** | 4 of 7 complete (5-7 skipped per task constraints: no smoke test stub, no commit, no TASK-INDEX) |
-| **Status** | completed (task work) — handoff to Wave 3 orchestrator |
-| **Next Action** | Hand back to Wave 3 orchestrator; task 055 (runtimeConfig consolidation) unblocked |
-| **Rigor Level** | FULL (auth tag, .ts modifications, ADR-028) |
-| **Knowledge loaded** | task POML 054, auth-init-divergence.md, task-053-factory-config-timing.md, createCodePageAuthInitializer.ts, both main.tsx, both authInit.ts, both runtimeConfig.ts, Spaarke.Auth/index.ts, DailyBriefing/authInit.ts (task 053 reference impl) |
-| **Files modified** | `src/solutions/LegalWorkspace/src/services/authInit.ts` (86→85 LOC, divergent logic → factory consumer); `src/solutions/SpaarkeAi/src/services/authInit.ts` (88→91 LOC, divergent logic → factory consumer); main.tsx in both solutions UNCHANGED (preserved consumer-import surface) |
-| **Pattern chosen** | Thin-wrapper lazy-singleton (task 053 pattern). LW has 22 consumers; SpaarkeAi has 1. Both use lazy because runtime config is `setRuntimeConfig`-after-`resolveRuntimeConfig` (NOT available at module load). Direct destructure would throw "Runtime config not initialized" at import time. |
-| **Build verification** | LegalWorkspace: `npm run build` → vite built in 13.29s, 2253 KB output (green). SpaarkeAi: surface-owned TS errors = 0; tsc-surface-gate ✓; Vite failure at `@spaarke/sdap-client` is pre-existing (documented in task 001 session note), unrelated to authInit migration. |
-| **Seam preservation** | SpaarkeAi `loadSpaarkeAiNotificationContext` is defined in `services/notificationContextLoader.ts` but NOT imported anywhere (task 002 deferred). No scaffolding to preserve in main.tsx; trivially satisfied by zero-touch. |
+| **Task** | 015 — Abstract dependencies; remove solution-local back-pointers (FR-07) |
+| **Step** | 6 of 6 task POML steps complete (step 7 TASK-INDEX update skipped per orchestrator: "NO commit. NO push. NO TASK-INDEX update") |
+| **Status** | completed (task work) — handoff to Wave 5 orchestrator |
+| **Next Action** | Hand back to Wave 5 orchestrator; task 016 (Pattern D registration shim) unblocked |
+| **Rigor Level** | FULL (refactoring tag, .ts modifications across 14 files, ADR-012 + ADR-024 + ADR-028) |
+| **Knowledge loaded** | task POML 015, project CLAUDE.md, project spec.md (FR-07), all 11 files in new package src/, source types/notifications.ts + notificationService.ts + preferencesService.ts + toastUtils.ts in solutions/DailyBriefing/, Spaarke.Auth/package.json (peer dep target), package.json of new package, sibling shared-lib tsconfigs (Smart Todo + Events for pattern parity) |
+| **Files hoisted (3 new in package)** | `src/client/shared/Spaarke.DailyBriefing.Components/src/types/notifications.ts` (323 LOC); `services/notificationService.ts` (330 LOC); `services/preferencesService.ts` (176 LOC); `utils/toastUtils.ts` (15 LOC) + `utils/index.ts` barrel |
+| **Files modified in new package** | `package.json` (added `@spaarke/auth` peer dep + `./utils` export); `src/index.ts` (added utils barrel re-export); `types/index.ts` (added `BriefingDependencies` interface + re-exported notifications surface); `services/index.ts` (added notification/preferences exports); `services/briefingService.ts` (`@spaarke/auth` direct import + intra-package `ChannelFetchResult`); 5 hooks (`useBriefingActions`, `useBriefingNarration`, `useBriefingNotifications`, `useBriefingPreferences`, `useInlineTodoCreate` — intra-package import paths); 3 components (`DailyBriefingApp`, `ActivityNotesSection`, `PreferencesDropdown` — intra-package import paths) |
+| **Shim files (4 in solutions/DailyBriefing)** | `types/notifications.ts` → `export * from "@spaarke/daily-briefing-components/types/notifications"`; `services/notificationService.ts` → re-export 5 funcs from `@spaarke/daily-briefing-components/services`; `services/preferencesService.ts` → re-export 2 funcs; `utils/toastUtils.ts` → `export { TOASTER_ID } from "@spaarke/daily-briefing-components/utils"` |
+| **package.json peer dep additions** | `@spaarke/auth: ^2.0.0` (per ADR-028; legit peer dep that task 010 scaffold should have included but didn't — orchestrator brief explicitly authorized this in task 015 scope) |
+| **`BriefingDependencies` interface (types/index.ts)** | `{ authenticatedFetch: AuthenticatedFetch, webApi: IWebApi, userId: string, tenantId: string, onRateLimitError?: (info) => void, onRecordOpen?: (info) => void }` — structural; concrete consumer wiring lives at standalone main.tsx + SpaarkeAi widget shell |
+| **Grep result** | `grep -rE "from\s+[\"'][^\"']*solutions/" src/client/shared/Spaarke.DailyBriefing.Components/src/` returns ZERO matches (clean). Also zero `Spaarke.Auth` back-pointer paths. All `solutions/` text in new package is JSDoc-comment-only (intentional hoist-origin history). |
+| **Build verification** | `npm run build` (`tsc --noEmit`) in new package returns 3 errors, all cross-package alias errors (`@spaarke/ui-components`, `@spaarke/ui-components/services`, `@spaarke/auth`) — same architectural pattern as sibling shared libs (Smart Todo + Events have identical `node` type-def errors when tsc-checked package-locally). Cross-package aliases resolve only at consumer site (Vite bundler resolution). ZERO new back-pointer errors. ZERO regressions in hoisted logic (byte-identical preservation). |
+| **Pattern chosen** | Hoist + re-export shim (Calendar + Smart Todo precedent). New package owns: types/notifications.ts, services/{notificationService,preferencesService,briefingService}.ts, utils/toastUtils.ts. Originals become 3-15 line shim files with `export * from "@spaarke/daily-briefing-components/…"`. Added `@spaarke/auth` peer dep + import path. Defined `BriefingDependencies` in types/index.ts. ADR-024 `TODO_REGARDING_CATALOG` in useInlineTodoCreate preserved verbatim (only `IWebApi`/`NotificationItem`/`NotificationPriority` type-import path rewritten). |
+| **Scope adherence** | Only files in `src/client/shared/Spaarke.DailyBriefing.Components/` + `src/solutions/DailyBriefing/` touched. NO BFF, LegalWorkspace, SpaarkeAi, authInit.ts, runtimeConfig.ts, MicrosoftToDoIcon, or TASK-INDEX.md touched (verified via git status). |
 
 ### Files Modified This Session
