+++ src/client/shared/Spaarke.DailyBriefing.Components/package.json
@@ -12,7 +12,8 @@
     "./widgets": "./src/widgets/index.ts",
     "./hooks": "./src/hooks/index.ts",
     "./services": "./src/services/index.ts",
-    "./types": "./src/types/index.ts"
+    "./types": "./src/types/index.ts",
+    "./utils": "./src/utils/index.ts"
   },
   "scripts": {
     "build": "tsc --noEmit",
@@ -21,6 +22,7 @@
   "peerDependencies": {
     "@fluentui/react-components": "^9.0.0",
     "@fluentui/react-icons": "^2.0.0",
+    "@spaarke/auth": "^2.0.0",
     "react": "^16.14.0 || ^17.0.0 || ^18.0.0 || ^19.0.0",
     "react-dom": "^16.14.0 || ^17.0.0 || ^18.0.0 || ^19.0.0"
   },
