+++ src/client/shared/Spaarke.DailyBriefing.Components/src/components/ActivityNotesSection.tsx
@@ -28,11 +28,11 @@ import {
   Skeleton,
   SkeletonItem,
 } from "@fluentui/react-components";
-import { CHANNEL_REGISTRY } from "../../../../../solutions/DailyBriefing/src/types/notifications";
+import { CHANNEL_REGISTRY } from "../types/notifications";
 import type {
   ChannelFetchResult,
   NotificationCategory,
-} from "../../../../../solutions/DailyBriefing/src/types/notifications";
+} from "../types/notifications";
 import { getChannelIcon } from "./channelIcons";
 import { ChannelHeading } from "./ChannelHeading";
 import { NarrativeBullet } from "./NarrativeBullet";
