+++ src/client/shared/Spaarke.DailyBriefing.Components/src/index.ts
@@ -23,4 +23,5 @@ export * from './components';
 export * from './widgets';
 export * from './hooks';
 export * from './services';
+export * from './utils';
 export type * from './types';
