+++ src/client/shared/Spaarke.DailyBriefing.Components/src/services/index.ts
@@ -9,6 +9,10 @@
  *
  * Populated by R2 task 012 (FR-09): hoisted `briefingService` (the BFF
  * `/summarize` + `/narrate` clients).
+ *
+ * Populated by R2 task 015 (FR-07): hoisted `notificationService` and
+ * `preferencesService` from the standalone DailyBriefing solution so the
+ * package no longer reaches back across the solution boundary.
  */
 
 export {
@@ -21,4 +25,17 @@ export {
   type TldrResult,
   type ChannelNarrationResult,
   type NarrativeBulletResult,
-} from './briefingService';
+} from "./briefingService";
+
+export {
+  fetchNotifications,
+  fetchAndGroupNotifications,
+  groupByCategory,
+  markNotificationRead,
+  markAllNotificationsRead,
+} from "./notificationService";
+
+export {
+  fetchDigestPreferences,
+  saveDigestPreferences,
+} from "./preferencesService";
