+++ src/client/shared/Spaarke.DailyBriefing.Components/src/utils/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @spaarke/daily-briefing-components — utils barrel
+ *
+ * Shared UI-side utility constants and helpers. R2 task 015 / FR-07 hoist
+ * (Toaster id constant moved here so the package no longer reaches back
+ * into the standalone DailyBriefing solution).
+ */
+
+export { TOASTER_ID } from "./toastUtils";
